api/requirements.txt
# Flask Framework
Flask==2.0.3
jinja2==3.1.1
flask_socketio==5.1.2
flask_cors==3.0.10
gevent==21.12.0


#goose lib and dependencies
# goose3==3.1.11
redis==4.3.1
psycopg2-binary==2.9.3

# Automated tests
pytest==3.0.5
pytest-cov==2.4.0
